requirements.txt
python-dotenv==1.0.0
vapi-server-sdk==1.5.0
assemblyai==0.17.0
python-Levenshtein==0.23.0
flask==2.3.2